New Implementation of Read and Write I2C
I studied the drivers found on this site

http://www.embedded-code.com/source-code/miscellaneous/real-time-clock-i
cs/maxim-ds1307-i2c-real-time-clock-driver

and rewrote it for our pic.

We had tried using it before, but turns out all of it's functions and
register names were  defined for versions OTHER THAN (I2C_V1) so I
changed all of them using find and replace and I checked to make sure
that the functions did the same thing by comparing the source files of
AckI2C() with AckI2C1() for example. They all did the exact same thing
but the registers had different names.